# test/e2e/__snapshots__/test_openapi/test_OpenAPIの形が変わっていないことを確認.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "Preset": {
        "description": "プリセット情報。",
        "properties": {
          "id": {
            "description": "プリセット ID 。",
            "title": "Id",
            "type": "integer"
          },
          "intonationScale": {
            "description": "選択した話者スタイルの感情表現の強弱を 0.0 ~ 2.0 の範囲で指定する (デフォルト: 1.0) 。\n「全体の抑揚」ではない点で VOICEVOX ENGINE と仕様が異なる。\n数値が大きいほど、選択した話者スタイルに近い感情表現が込められた声になる。\n例えば話者スタイルが「上機嫌」なら、数値が大きいほどより嬉しそうな明るい話し方になる。\n一方で、話者やスタイルによっては、数値を上げすぎると発声がおかしくなったり、棒読みで不自然な声になる場合もある。\n正しく発声できる上限値は話者やスタイルごとに異なる。必要に応じて最適な値を見つけて調整すること。\n全スタイルの平均であるノーマルスタイルでは自動で適切な感情表現が選択されるため、この値を指定しても無視される。",
            "title": "Intonationscale",
            "type": "number"
          },
          "name": {
            "description": "プリセット名。",
            "title": "Name",
            "type": "string"
          },
          "pauseLength": {
            "description": "句読点などの無音時間。null のときは無視される。デフォルト値は null 。",
            "title": "AivisSpeech Engine ではサポートされていないフィールドです (常に無視されます)",
            "type": "number"
          },
          "pauseLengthScale": {
            "default": 1,
            "description": "句読点などの無音時間（倍率）。デフォルト値は 1 。",
            "title": "AivisSpeech Engine ではサポートされていないフィールドです (常に無視されます)",
            "type": "number"
          },
          "pitchScale": {
            "description": "全体の音高を -0.15 ~ 0.15 の範囲で指定する (デフォルト: 0.0) 。\n数値が大きいほど高い声になる。\nVOICEVOX ENGINE と異なり、この値を 0.0 から変更すると音質が劣化するため注意が必要。",
            "title": "Pitchscale",
            "type": "number"
          },
          "postPhonemeLength": {
            "description": "音声の後の無音時間 (秒)。",
            "title": "Postphonemelength",
            "type": "number"
          },
          "prePhonemeLength": {
            "description": "音声の前の無音時間 (秒)。",
            "title": "Prephonemelength",
            "type": "number"
          },
          "speaker_uuid": {
            "description": "キャラクターの UUID 。",
            "title": "Speaker Uuid",
            "type": "string"
          },
          "speedScale": {
            "description": "全体の話速を 0.5 ~ 2.0 の範囲で指定する (デフォルト: 1.0) 。\n2.0 で 2 倍速、0.5 で 0.5 倍速になる。",
            "title": "Speedscale",
            "type": "number"
          },
          "style_id": {
            "description": "スタイル ID 。",
            "title": "Style Id",
            "type": "integer"
          },
          "tempoDynamicsScale": {
            "default": 1.0,
            "description": "話す速さ（テンポ）の緩急の強弱を 0.0 ~ 2.0 の範囲で指定する (デフォルト: 1.0) 。\nAivisSpeech Engine 固有のフィールドで、VOICEVOX ENGINE には存在しない。\n数値が大きいほどより早口で生っぽい抑揚がついた声になる。\nVOICEVOX ENGINE との互換性のため、未指定時はデフォルト値が適用される。",
            "title": "Tempodynamicsscale",
            "type": "number"
          },
          "volumeScale": {
            "description": "全体の音量を 0.0 ~ 2.0 の範囲で指定する (デフォルト: 1.0) 。\n数値が大きいほど大きな声になる。",
            "title": "Volumescale",
            "type": "number"
          }
        },
        "required": [
          "id",
          "name",
          "speaker_uuid",
          "style_id",
          "speedScale",
          "intonationScale",
          "pitchScale",
          "volumeScale",
          "prePhonemeLength",
          "postPhonemeLength"
        ],
        "title": "Preset",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "AivisSpeech の音声合成エンジンです。",
    "title": "AivisSpeech Engine",
    "version": "latest"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/add_preset": {
      "post": {
        "description": "新しいプリセットを追加します。",
        "operationId": "add_preset",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Preset",
                "description": "新しいプリセット。プリセットIDが既存のものと重複している場合は、新規のプリセットIDが採番されます。"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Add Preset Add Preset Post",
                  "type": "integer"
                }
              }
            },
            "description": "追加したプリセットのプリセットID"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "新しいプリセットを追加する",
        "tags": [
          "プリセット"
        ]
      }
    },
...
    "/delete_preset": {
      "post": {
        "description": "既存のプリセットを削除します。",
        "operationId": "delete_preset",
        "parameters": [
          {
            "description": "削除するプリセットのプリセットID",
            "in": "query",
            "name": "id",
            "required": true,
            "schema": {
              "description": "削除するプリセットのプリセットID",
              "title": "Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "既存のプリセットを削除する",
        "tags": [
          "プリセット"
        ]
      }
    },
...
    "/update_preset": {
      "post": {
        "description": "既存のプリセットを更新します。",
        "operationId": "update_preset",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Preset",
                "description": "更新するプリセット。プリセットIDが更新対象と一致している必要があります。"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Update Preset Update Preset Post",
                  "type": "integer"
                }
              }
            },
            "description": "更新したプリセットのプリセットID"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "既存のプリセットを更新する",
        "tags": [
          "プリセット"
        ]
      }
    },
...
  }
}